refactor(sdk): split geniex-bench into per-concern modules
Pure code movement, no behavior change: benchmark.c had grown to 2174
lines mixing argv parsing, model-manager resolution, run loops,
aggregation and report writing. Every function body moves verbatim —
bench.h carries the type definitions and a map of which module owns what.

The only edits to moved code are the two globals that now cross a module
boundary: g_token_callback_delay_us gets a set_token_callback_delay_us()
setter for parse_args to call, and main()'s g_mm_inited teardown becomes
mm_shutdown(). Signatures that lost `static` are re-wrapped by
clang-format.
